fix(p2): match_pulse_levels 按 start_row 裁剪后再转 float，修复真实 BTS 头部行转换回归
P2 将 battery_analysis 的 .tolist() 改为 .to_numpy(dtype=float) 后，
真实 BTS 导出文件头部行（第 0 行测试名、第 1 行 '电流(A)'/'电压(V)' 列名）
被全数组强转 float 抛 "could not convert string to float: '电流(A)'"，
导致文件被跳过。旧实现从 start_row=2 起逐行 float() 天然跳过头部。

修复：pulse_matcher 内部先按 start_row 裁剪再做 dtype=float 转换，
返回行索引 +start 补偿为原始绝对索引；battery_analysis 调用处改为
to_numpy() 不提前强转。新增真实头部行回归测试。

验证：真实文件（48.7 万行/27.4 万脉冲行）完整 _parallel_process_file
走通；551 passed / 9 skipped；benchmark 不退化（match 0.0097s、charge batch 0.0023s）。

+++ src/battery_analysis/utils/processors/battery_analysis.py
@@ -296,9 +296,12 @@ class BatteryAnalysis:
             raise BatteryAnalysisException(f"Pulse data not found: {strPath}")
 
         # ── 脉冲等级匹配 ────────────────────────────────────────
+        # 不在此处 to_numpy(dtype=float)：record 列含头部标题行（行1 '电流(A)' 等
+        # 非数值单元格），全数组强转会抛 ValueError；由 match_pulse_levels 按
+        # start_row=2 裁剪后再转换，等价旧实现从 start_row 起逐行 float()。
         matched = match_pulse_levels(
-            record_current.to_numpy(dtype=float),
-            record_voltage.to_numpy(dtype=float),
+            record_current.to_numpy(),
+            record_voltage.to_numpy(),
             pulse_mask.to_numpy(dtype=bool),
             listCurrentLevel,
             listVoltageLevel,
